restructure shit
